norm and mem leaks

+++ src/lexer.c
@@ -45,7 +45,27 @@ int	check_pipes(t_token *tokens, t_env *env)
 	}
 	if (!flag)
 		return (0);
+	ft_putstr_fd("Incorrect pipes\n", STDERR_FILENO);
 	return (1);
+}
+
+int	if_not_space(t_token *tokens)
+{
+	int		i;
+	bool	flag;
+
+	i = 0;
+	flag = false;
+	while (tokens)
+	{
+		if (tokens->type == SEPERATOR)
+			flag = true;
+		i++;
+		tokens = tokens->next;
+	}
+	if (flag && i == 1)
+		return (1);
+	return (0);
 }
 
 void	lexer(char *line, t_env *my_env)
@@ -58,15 +78,14 @@ void	lexer(char *line, t_env *my_env)
 	{
 		tokenize(line, &tokens);
 		expander(&tokens, my_env);
-		if (tokens && !check_pipes(tokens, my_env))
+		if (tokens && !if_not_space(tokens) && !check_pipes(tokens, my_env))
 		{
 			commands = merge_tokens(tokens, my_env);
-			// print_cmds(commands);
 			if (commands)
 				run_commands(commands, my_env);
-			tokens = destroy_tokens(tokens);
 			commands = destroy_cmds(commands);
 		}
+		tokens = destroy_tokens(tokens);
 	}
 	else
 	{
